Input crate: code re-organization, added some impls
